firmware & co: load of improvements
-Fixed an issue with NFC module that would prevent receiving data
from module due to not reading from the same thread we opened the
serial port
-Fixed services
-Added property to device-state document to specify if the document
is partial or not. Partial document allow one to update partially the state of a device
that would merge it to its complete state, ollowing for 'cmd' like documents.
For exemple, you can use a docuement to change juste the 'on' state of a light.
Once the bulb have updated its internal state accordingly, it should republish is full state
in order for subscribers to update properly.
Thus, a subscriber shall subscribe for document where 'partial_state' is False:

{
  and : [
    {
      equals: {
        device_id: 'device id'
      }
    },
    {
      equals: {
        partial_state: false
      }
    }
  ]
}
age for your changes. Lines starting